requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
numpy==2.3.0
pandas==2.3.0
pillow==11.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
